refactor(trading-signals): extract zero-cross signal into shared base class
Eight oscillators (AO, TRIX, MOM, PPO, ADOSC, CMF, EMV, VROC) carried
byte-identical zero-line signal logic in two stylistic variants. The new
ZeroCrossTrendIndicatorSeries encodes the interpretation once: positive
readings signal bullish pressure, negative bearish, exactly zero sideways.

AC keeps its own implementation on purpose — it maps zero to BEARISH and
has no sideways state.
